requirement.txt
aioredis==1.3.1
amqp==2.5.2
ansible==2.9.4
anyjson==0.3.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
billiard==3.6.1.0
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.41.0
django-appconf==1.0.3
django-celery==3.3.1
django-celery-beat==1.6.0
django-celery-results==1.1.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-imagekit==4.0.2
django-ratelimit==2.0.0
django-redis==4.11.0
django-redis-sessions==0.6.1
django-timezone-field==3.1
dnspython==1.16.0
eventlet==0.25.1
future==0.18.2
google-auth==1.10.1
greenlet==0.4.15
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.4.0
incremental==17.5.0
Jinja2==2.11.0
kombu==4.6.7
kubernetes==10.0.1
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.1.0
msgpack==0.6.2
msgpack-python==0.5.6
multi-key-dict==2.0.3
mysqlclient==1.4.6
oauthlib==3.1.0
pbr==5.4.4
pilkit==2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyguacamole==0.9
PyHamcrest==2.0.0
pylibmc==1.6.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-jenkins==1.6.0
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.7
vine==1.3.0
websocket-client==0.57.0
zipp==1.0.0
zope.interface==4.7.1
